geolocation/geo-00062.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: watchPosition returns synchronously before any callbacks are invoked</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#watch-position">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var t =async_test("Test checks that watchPosition returns synchronously before any callbacks are invoked");
      setup({timeout: 20000});
      try {
        var hasMethodReturned = false;
        window.navigator.geolocation.watchPosition(checkMethodHasReturned, checkMethodHasReturned);
        hasMethodReturned = true;
      } catch (err) {
        t.step(function () {
          assert_unreached("Error message: " + err.message);
        });
        t.done();
      }
      function checkMethodHasReturned() {
        t.step(function () {
          assert_true(hasMethodReturned, "watchPosition returns synchronously before any callbacks are invoked")
        });
        t.done();
      }
    </script>
  </body>
</html>
